# bsin-server-apps/bsin-server-ai/ai-go-wechat/docs/swagger.yaml
host: localhost:8126
info:
  contact:
    email: leijiwu001@gmail.com
    name: leonard
    url: https://www.s11edao.com
  description: bsin-paas golang 后台应用
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: bsin-paas golang server
  version: "1.0"
paths:
...
  /account/createUser:
    post:
      description: create account
      operationId: "11"
      parameters:
      - description: 签名信息:暂时无用，可以随便写死一个
        in: body
        name: sign
        required: true
        schema:
          type: string
      - description: 用户设置的交易密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"retCode":200,"account":"用户的链上账户","privateKey":"用户转账nft需要用到的私钥","retMsg":"ok"}'
          schema:
            type: string
        "400":
          description: '{"retCode":500,"retMsg":"failed"}'
          schema:
            type: string
      summary: 创建账户
      tags:
      - blockchain/cfx/account
...
swagger: "2.0"